AI_LearnWords/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
cachetools==5.3.2
certifi==2024.2.2
click==8.1.7
fastapi==0.110.0
fastapi_poe==0.0.34
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
httpx-sse==0.4.0
idna==3.6
pydantic==2.6.2
pydantic_core==2.16.3
PyJWT==2.8.0
python-dotenv==1.0.1
sniffio==1.3.1
sse-starlette==2.0.0
starlette==0.36.3
typing_extensions==4.10.0
uvicorn==0.27.1
zhipuai==2.0.1
